adding folder disks window
The Folder Disks window is a modeless dialog that can be opened as a companion to the main window. The UI was conceived out of the need to easily navigate an archived set of floppy disks placed in a single folder. Thumbnail images (currently an identically named *.png file) can also be placed in the folder to provide a screenshot representing the main content of any particular disk image.